% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_helpers.R
\name{nuke}
\alias{nuke}
\title{Replace all occurrences of a value in a data frame.}
\usage{
nuke(df, nuke_value = NA, ash = 0, regex = FALSE)
}
\arguments{
\item{df}{The data frame to which you're replacing values.}

\item{nuke_value}{The value you want to replace with \code{ash}.}

\item{ash}{The value you want to replace \code{nuke_value} with.}

\item{regex}{Logical indicating whether or not you want to use \code{grepl(ignore.case = TRUE, perl = TRUE)} style regex.}
}
\value{
A data frame with the all occurrences of \code{nuke_value} replaced with \code{ash}.
}
\description{
Replace all occurrences of a value in a data frame.
}
\examples{
nuke(band_members, "Mick", "IT WORKED")
nuke(mtcars, 6, "IT WORKED AGAIN")
nuke(mtcars, 6)
}
